+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - use serde with derive feature instead of serde_derive, fmt&clippy 
 - prevent panic if step index out of bounds
+- don't send Network.clearBrowserCache and Network.clearBrowserCookies
 
 ## [0.1.1] - 2020-01-14